refactor: use module.registerHooks for the gi: ESM loader
Switch the `gi:` scheme loader from `module.register` (async hooks on a
separate loader thread, loaded by URL) to `module.registerHooks` (synchronous,
in-thread hooks passed directly as functions).

Since registerHooks takes the hook functions rather than a module URL, the
whole implementation now lives in register.mjs and the separate hooks.mjs file
(plus its `./hooks` package export, which was unused) is removed. The resolve
and load hooks are now synchronous, as registerHooks requires; the synthetic
`gi.require` module emitted by `load` is unchanged.

Bumps the minimum from Node >= 20.6 (module.register) to Node >= 22.15
(module.registerHooks); node-gtk v4 already targets Node 22/24/26.
