+++ package.json
@@ -9,7 +9,6 @@
       "require": "./lib/index.js"
     },
     "./register": "./lib/esm/register.mjs",
-    "./hooks": "./lib/esm/hooks.mjs",
     "./styles": {
       "types": "./lib/styles.d.ts",
       "default": "./lib/styles.js"
